+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Next Version
 ### Improvements
  * Exceptions when collecting suggestions for autocompletion no longer crash the entire application.
+ * JSON array schemas can now specify a `minCount` and a `maxCount`.
 
 ### Fixes
  * Fixed crash when attempting to load project with a datapack version that is not supported.